+++ cextern/wcslib/C/flexed/fitshdr.c
@@ -10085,8 +10085,8 @@ char *fitshdrtext;
 #line 1 "fitshdr.l"
 /*============================================================================
 
-  WCSLIB 5.14 - an implementation of the FITS WCS standard.
-  Copyright (C) 1995-2016, Mark Calabretta
+  WCSLIB 5.16 - an implementation of the FITS WCS standard.
+  Copyright (C) 1995-2017, Mark Calabretta
 
   This file is part of WCSLIB.
 
@@ -10107,7 +10107,7 @@ char *fitshdrtext;
 
   Author: Mark Calabretta, Australia Telescope National Facility, CSIRO.
   http://www.atnf.csiro.au/people/Mark.Calabretta
-  $Id: fitshdr.c,v 5.14 2016/02/07 10:49:31 mcalabre Exp $
+  $Id: fitshdr.c,v 5.16 2017/01/15 04:25:01 mcalabre Exp $
 *=============================================================================
 *
 * fitshdr.l is a Flex description file containing a lexical scanner
@@ -10274,7 +10274,7 @@ static int input (void );
 #ifndef YY_INPUT
 #define YY_INPUT(buf,result,max_size) \
 	errno=0; \
-	while ( (result = read( fileno(fitshdrin), (char *) buf, max_size )) < 0 ) \
+	while ( (result = read( fileno(fitshdrin), (char *) buf, (yy_size_t) max_size )) < 0 ) \
 	{ \
 		if( errno != EINTR) \
 		{ \
@@ -11151,7 +11151,7 @@ static int yy_get_next_buffer (void)
 
 	else
 		{
-			yy_size_t num_to_read =
+			int num_to_read =
 			YY_CURRENT_BUFFER_LVALUE->yy_buf_size - number_to_move - 1;
 
 		while ( num_to_read <= 0 )